refactor: expose internal packages
