+++ go.mod
@@ -1,10 +1,8 @@
-module github.com/autonomy/gitmeta
+module github.com/talos-systems/gitmeta
 
 require (
 	github.com/Masterminds/semver v1.4.2
-	github.com/inconshreveable/mousetrap v1.0.0 // indirect
 	github.com/spf13/cobra v0.0.3
 	github.com/spf13/pflag v1.0.3
-	gopkg.in/src-d/go-billy.v4 v4.3.0 // indirect
 	gopkg.in/src-d/go-git.v4 v4.8.1
 )
